Altered Scanner Beams
- Changed collision boxes in scanner beam to be much smaller to account
for the fact the circle collider on the player is bigger than it. This
fixes the problem of getting caught by the laser when you are only close
to it.

- Changed the Z-Axis of all scanner beams so on the full build the
scanner beam is behind the map. This fixes the problem on a build where
the scanner beam would show up over the map which did not look right.